aggregator-service/requirements.txt
Flask
pymysql
mysql-connector-python
APScheduler